requirements.txt
flask==3.0.0
flask-pymongo==2.3.0
python-dotenv==1.0.0
pytest==7.4.3
mongomock==4.1.2